+++ Session/CookieSessionHandler.php
@@ -13,10 +13,15 @@ namespace Nelmio\SecurityBundle\Session;
 
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
+use Symfony\Component\HttpKernel\Event\ResponseEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpFoundation\Cookie;
 use Psr\Log\LoggerInterface;
 
+/**
+ * @final
+ */
 class CookieSessionHandler implements \SessionHandlerInterface
 {
     protected $request;
@@ -58,10 +63,15 @@ class CookieSessionHandler implements \SessionHandlerInterface
     }
 
     /**
-     * @param FilterResponseEvent $e
+     * @param FilterResponseEvent|ResponseEvent $e
      */
-    public function onKernelResponse(FilterResponseEvent $e)
+    public function onKernelResponse($e)
     {
+        // Compatibility with Symfony < 5 and Symfony >=5
+        if (!$e instanceof FilterResponseEvent && !$e instanceof ResponseEvent) {
+            throw new \InvalidArgumentException(\sprintf('Expected instance of type %s, %s given', \class_exists(ResponseEvent::class) ? ResponseEvent::class : FilterResponseEvent::class, \is_object($e) ? \get_class($e) : \gettype($e)));
+        }
+
         if (HttpKernelInterface::MASTER_REQUEST !== $e->getRequestType()) {
             return;
         }
@@ -91,10 +101,15 @@ class CookieSessionHandler implements \SessionHandlerInterface
     }
 
     /**
-     * @param GetResponseEvent $e
+     * @param GetResponseEvent|RequestEvent $e
      */
-    public function onKernelRequest(GetResponseEvent $e)
+    public function onKernelRequest($e)
     {
+        // Compatibility with Symfony < 5 and Symfony >=5
+        if (!$e instanceof GetResponseEvent && !$e instanceof RequestEvent) {
+            throw new \InvalidArgumentException(\sprintf('Expected instance of type %s, %s given', \class_exists(RequestEvent::class) ? RequestEvent::class : GetResponseEvent::class, \is_object($e) ? \get_class($e) : \gettype($e)));
+        }
+
         if (HttpKernelInterface::MASTER_REQUEST !== $e->getRequestType()) {
             return;
         }