+++ Tests/Listener/ClickjackingListenerTest.php
@@ -15,7 +15,6 @@ use Nelmio\SecurityBundle\EventListener\ClickjackingListener;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
-use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 
 class ClickjackingListenerTest extends \PHPUnit\Framework\TestCase
 {
@@ -64,13 +63,35 @@ class ClickjackingListenerTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals(null, $response->headers->get('X-Frame-Options'));
     }
 
+    public function testWrongEventClass()
+    {
+        $this->expectException(\InvalidArgumentException::class);
+
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $this->expectExceptionMessage('Expected instance of type Symfony\Component\HttpKernel\Event\ResponseEvent, Symfony\Component\HttpFoundation\Response given');
+        } else {
+            $this->expectExceptionMessage('Expected instance of type Symfony\Component\HttpKernel\Event\FilterResponseEvent, Symfony\Component\HttpFoundation\Response given');
+        }
+
+        $response = new Response();
+        $this->listener->onKernelResponse($response);
+
+        return $response;
+    }
+
     protected function callListener($listener, $path, $masterReq, $contentType = 'text/html')
     {
         $request = Request::create($path);
         $response = new Response();
         $response->headers->add(array('content-type' => $contentType));
 
-        $event = new FilterResponseEvent($this->kernel, $request, $masterReq ? HttpKernelInterface::MASTER_REQUEST : HttpKernelInterface::SUB_REQUEST, $response);
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $event = new $class($this->kernel, $request, $masterReq ? HttpKernelInterface::MASTER_REQUEST : HttpKernelInterface::SUB_REQUEST, $response);
         $listener->onKernelResponse($event);
 
         return $response;