+++ Tests/Listener/EncryptedCookieListenerTest.php
@@ -17,8 +17,6 @@ use Symfony\Component\HttpFoundation\Cookie;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
-use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
-use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 
 class EncryptedCookieListenerTest extends \PHPUnit\Framework\TestCase
 {
@@ -51,7 +49,13 @@ class EncryptedCookieListenerTest extends \PHPUnit\Framework\TestCase
         $listener = new EncryptedCookieListener($this->encrypter, $encryptedCookieNames);
         $request = Request::create('/', 'GET', array(), $inputCookies);
 
-        $event = new GetResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST);
+        if (class_exists('Symfony\Component\HttpKernel\Event\RequestEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\RequestEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\GetResponseEvent';
+        }
+
+        $event = new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST);
         $listener->onKernelRequest($event);
 
         $this->assertSame($expectedCookies, $request->cookies->all());
@@ -82,7 +86,13 @@ class EncryptedCookieListenerTest extends \PHPUnit\Framework\TestCase
             $response->headers->setCookie(method_exists('Symfony\\Component\\HttpFoundation\\Cookie', 'create') ? Cookie::create($name, $cookie) : new Cookie($name, $cookie));
         }
 
-        $event = new FilterResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response);
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $event = new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response);
         $listener->onKernelResponse($event);
 
         $responseCookieValues = array();
@@ -120,7 +130,13 @@ class EncryptedCookieListenerTest extends \PHPUnit\Framework\TestCase
             $response->headers->setCookie(method_exists('Symfony\\Component\\HttpFoundation\\Cookie', 'create') ? Cookie::create($name, $cookie) : new Cookie($name, $cookie));
         }
 
-        $event = new FilterResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response);
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $event = new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response);
         $listener->onKernelResponse($event);
 
         $responseCookieValues = array();
@@ -132,7 +148,13 @@ class EncryptedCookieListenerTest extends \PHPUnit\Framework\TestCase
 
         $request = Request::create('/', 'GET', array(), $responseCookieValues);
 
-        $event = new GetResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST);
+        if (class_exists('Symfony\Component\HttpKernel\Event\RequestEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\RequestEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\GetResponseEvent';
+        }
+
+        $event = new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST);
         $listener->onKernelRequest($event);
 
         $this->assertSame($inputCookies, $request->cookies->all());