+++ Tests/Listener/XssProtectionListenerTest.php
@@ -5,7 +5,6 @@ namespace Nelmio\SecurityBundle\Tests\Listener;
 use Nelmio\SecurityBundle\EventListener\XssProtectionListener;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 
 class XssProtectionListenerTest extends \PHPUnit\Framework\TestCase
@@ -43,7 +42,13 @@ class XssProtectionListenerTest extends \PHPUnit\Framework\TestCase
         $request = Request::create($path);
         $response = new Response();
 
-        $event = new FilterResponseEvent(
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $event = new $class(
             $this->kernel,
             $request,
             $masterReq ? HttpKernelInterface::MASTER_REQUEST : HttpKernelInterface::SUB_REQUEST,