+++ Tests/Session/CookieSessionHandlerTest.php
@@ -14,8 +14,6 @@ namespace Nelmio\SecurityBundle\Tests\Session;
 use Nelmio\SecurityBundle\Session\CookieSessionHandler;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Event\GetResponseEvent;
-use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 
 class CookieSessionHandlerTest extends \PHPUnit\Framework\TestCase
@@ -30,19 +28,17 @@ class CookieSessionHandlerTest extends \PHPUnit\Framework\TestCase
         $this->kernel = $this->getMockBuilder('Symfony\Component\HttpKernel\HttpKernelInterface')->getMock();
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testOpenWithNoRequest()
     {
+        $this->expectException('RuntimeException');
+
         $this->handler->open('foo', 'bar');
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testReadWithNoRequest()
     {
+        $this->expectException('RuntimeException');
+
         $this->handler->read('foo');
     }
 
@@ -54,12 +50,25 @@ class CookieSessionHandlerTest extends \PHPUnit\Framework\TestCase
         $session->expects($this->exactly(1))->method('save');
         $request->setSession($session);
 
-        $this->handler->onKernelRequest(new GetResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST));
+        if (class_exists('Symfony\Component\HttpKernel\Event\RequestEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\RequestEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\GetResponseEvent';
+        }
+
+        $this->handler->onKernelRequest(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST));
 
         $this->assertTrue($this->handler->open('foo', 'bar'));
 
         $this->handler->write('sessionId', 'mydata');
-        $this->handler->onKernelResponse(new FilterResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
+
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $this->handler->onKernelResponse(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
 
         $cookies = $response->headers->getCookies();
 
@@ -78,8 +87,21 @@ class CookieSessionHandlerTest extends \PHPUnit\Framework\TestCase
         $session->expects($this->exactly(2))->method('save');
         $request->setSession($session);
 
-        $this->handler->onKernelRequest(new GetResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST));
-        $this->handler->onKernelResponse(new FilterResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
+        if (class_exists('Symfony\Component\HttpKernel\Event\RequestEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\RequestEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\GetResponseEvent';
+        }
+
+        $this->handler->onKernelRequest(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST));
+
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $this->handler->onKernelResponse(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
 
         $cookies = $response->headers->getCookies();
 
@@ -89,7 +111,13 @@ class CookieSessionHandlerTest extends \PHPUnit\Framework\TestCase
 
         $this->handler->destroy('sessionId');
 
-        $this->handler->onKernelResponse(new FilterResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $this->handler->onKernelResponse(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
 
         $cookies = $response->headers->getCookies();
 
@@ -117,7 +145,20 @@ class CookieSessionHandlerTest extends \PHPUnit\Framework\TestCase
         $request->setSession($session);
         $response->headers = $headers;
 
-        $this->handler->onKernelRequest(new GetResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST));
-        $this->handler->onKernelResponse(new FilterResponseEvent($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
+        if (class_exists('Symfony\Component\HttpKernel\Event\RequestEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\RequestEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\GetResponseEvent';
+        }
+
+        $this->handler->onKernelRequest(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST));
+
+        if (class_exists('Symfony\Component\HttpKernel\Event\ResponseEvent')) {
+            $class = 'Symfony\Component\HttpKernel\Event\ResponseEvent';
+        } else {
+            $class = 'Symfony\Component\HttpKernel\Event\FilterResponseEvent';
+        }
+
+        $this->handler->onKernelResponse(new $class($this->kernel, $request, HttpKernelInterface::MASTER_REQUEST, $response));
     }
 }